refactor: remove folders
